+++ src/app/terra-components.module.ts
@@ -145,6 +145,7 @@ import { TerraMultiSplitViewBreadcrumbsService } from './components/split-view/m
 import { RouterModule } from '@angular/router';
 import { TerraTimePickerComponentExample } from './components/forms/input/time-picker/example/terra-time-picker.component.example';
 import { TerraTwoColumnsContainerComponent } from './components/layouts/column-container/terra-two-columns-container.component';
+import { TerraTwoColumnsContainerDirective } from './components/layouts/column-container/terra-two-columns-container.directive';
 import { TerraBreadcrumbsComponent } from './components/breadcrumbs/terra-breadcrumbs.component';
 import { TerraBreadcrumbsService } from './components/breadcrumbs/service/terra-breadcrumbs.service';
 import { Type } from '@angular/core/src/type';
@@ -218,6 +219,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraDynamicSwitchComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
+        TerraTwoColumnsContainerDirective,
         TerraBreadcrumbsComponent,
 
         // ### declarations of terra-components examples
@@ -376,6 +378,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraDynamicSwitchComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
+        TerraTwoColumnsContainerDirective,
         TerraBreadcrumbsComponent,
 
         // ### declarations of terra-components examples